Tiles setup
Tiles configuration setup
Sample URL http://localhost:8080/2let_war_exploded/item
