--- app/src/main/java/edu/mum/cs/waa/DeploymentDescriptor.java
@@ -1,37 +1,0 @@
-package edu.mum.cs.waa;
-
-import org.springframework.web.WebApplicationInitializer;
-import org.springframework.web.context.ContextLoaderListener;
-import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
-import org.springframework.web.filter.DelegatingFilterProxy;
-import org.springframework.web.servlet.DispatcherServlet;
-
-import javax.servlet.ServletContext;
-import javax.servlet.ServletRegistration;
-
-public class DeploymentDescriptor implements WebApplicationInitializer {
-
-	@Override
-	public void onStartup(ServletContext container) {
-		// Create the 'root' Spring application context
-		AnnotationConfigWebApplicationContext rootContext = new AnnotationConfigWebApplicationContext();
-		rootContext.register(Service.class, Persistence.class, Security.class);
-
-		// Manage the lifecycle of the root application context
-		container.addListener(new ContextLoaderListener(rootContext));
-
-		// Create the dispatcher servlet's Spring application context
-		AnnotationConfigWebApplicationContext dispatcherServlet = new AnnotationConfigWebApplicationContext();
-		dispatcherServlet.register(Dispatcher.class);
-
-		// Register and map the dispatcher servlet
-		ServletRegistration.Dynamic dispatcher = container.addServlet("dispatcher",
-				new DispatcherServlet(dispatcherServlet));
-		dispatcher.setLoadOnStartup(1);
-		dispatcher.addMapping("/");
-
-		container.addFilter("springSecurityFilterChain", new DelegatingFilterProxy("springSecurityFilterChain"))
-				.addMappingForUrlPatterns(null, false, "/*");
-	}
-
-}
