--- app/src/main/java/edu/mum/cs/waa/Dispatcher.java
@@ -1,69 +1,0 @@
-package edu.mum.cs.waa;
-
-import edu.mum.cs.waa.interceptor.NoticeInterceptor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
-import org.springframework.context.support.ReloadableResourceBundleMessageSource;
-import org.springframework.core.env.Environment;
-import org.springframework.validation.Validator;
-import org.springframework.validation.beanvalidation.LocalValidatorFactoryBean;
-import org.springframework.web.servlet.config.annotation.*;
-import org.springframework.web.servlet.view.InternalResourceViewResolver;
-
-@EnableWebMvc
-@Configuration
-@ComponentScan(basePackages = { "edu.mum.cs.waa.controller" })
-@PropertySource(value = "classpath:application.properties")
-public class Dispatcher extends WebMvcConfigurerAdapter {
-
-	@Autowired
-	Environment environment;
-
-	@Override
-	public void addResourceHandlers(ResourceHandlerRegistry registry) {
-		registry.addResourceHandler(environment.getProperty("staticResourceAlias"))
-				.addResourceLocations(environment.getProperty("staticResourceLocation"));
-	}
-
-	@Override
-	public void configureDefaultServletHandling(DefaultServletHandlerConfigurer configurer) {
-		configurer.enable();
-	}
-
-	@Bean
-	public InternalResourceViewResolver jspViewResolver() {
-		InternalResourceViewResolver bean = new InternalResourceViewResolver();
-		bean.setPrefix(environment.getProperty("viewResolver.prefix"));
-		bean.setSuffix(environment.getProperty("viewResolver.suffix"));
-		return bean;
-	}
-
-	@Override
-	public void addInterceptors(InterceptorRegistry registry) {
-		registry.addInterceptor(new NoticeInterceptor()).addPathPatterns("/employees/list");
-	}
-
-	@Bean(name = "messageSource")
-	public ReloadableResourceBundleMessageSource messageSource() {
-		ReloadableResourceBundleMessageSource resource = new ReloadableResourceBundleMessageSource();
-		resource.setBasename("classpath:messages");
-		resource.setDefaultEncoding("UTF-8");
-		return resource;
-	}
-
-	@Bean(name = "validator")
-	public LocalValidatorFactoryBean validator() {
-		LocalValidatorFactoryBean bean = new LocalValidatorFactoryBean();
-		bean.setValidationMessageSource(messageSource());
-		return bean;
-	}
-
-	@Override
-	public Validator getValidator() {
-		return validator();
-	}
-
-}
