+++ app/src/main/webapp/WEB-INF/views/pages/item.jsp
@@ -1,0 +1,8 @@
+<%--
+  Created by IntelliJ IDEA.
+  User: Duncan
+  Date: 20/11/2019
+  Time: 14:56
+  To change this template use File | Settings | File Templates.
+--%>
+<h2>Item Page</h2>
