+++ app/src/main/webapp/WEB-INF/views/tiles/template/defaultFooter.jsp
@@ -1,0 +1,8 @@
+<%--
+  Created by IntelliJ IDEA.
+  User: Duncan
+  Date: 20/11/2019
+  Time: 14:48
+  To change this template use File | Settings | File Templates.
+--%>
+<p>Footer</p>
