+++ app/src/main/webapp/WEB-INF/views/tiles/template/defaultMenu.jsp
@@ -1,0 +1,14 @@
+<%--
+  Created by IntelliJ IDEA.
+  User: Duncan
+  Date: 20/11/2019
+  Time: 14:51
+  To change this template use File | Settings | File Templates.
+--%>
+<nav>
+    <ul id="menu">
+        <li><a href="${pageContext.request.contextPath}/">Home</a></li>
+        <li><a href="${pageContext.request.contextPath}/products">Products</a></li>
+        <li><a href="${pageContext.request.contextPath}/contactus">Contact Us</a></li>
+    </ul>
+</nav>
